Car now does PID and speed scaling correctly. Runs smootly at 45.

+++ pwm.h
@@ -4,7 +4,7 @@
 void SetDutyCycle(unsigned int leftDuty, unsigned int rightDuty, unsigned int Frequency, int dir);
 void InitPWM();
 void PWM_ISR();
-void SetServoDutyCycle(unsigned int DutyCycle, unsigned int Frequency, int dir);
+void SetServoDutyCycle(float DutyCycle, unsigned int Frequency, int dir);
 void InitServoPWM();
 
 #endif /* PWM_H_ */
